Add Jupyter notebook for calculating a prior distribution
on vechicle ages.

Next: Account for the fact that it is only March in the current year
(making it less likely to see vehicles of age 0 years).

Also, we are not accounting for supply and demand.  Just because there
are a lot of 2-year-old cars that have survived, doesn't mean that a lot
of people will own one.