Try some more kriging stuff but using yr_diff.
Next: Just do kriging for each model_year separately because all that we
care about is using it for interpolation.  Alternatives to kriging can also
be pursued.  Then, multiply predicted advertised inventory
counts by an affine function for the other model years.  The parameters
for the affine function can be chosen so that they minimize the
mean squared error.  Constraints on the parameters of the affine
function can be set up so that any roots of it are outside the interval
[-1.5, 1.5] (for the yr_diff = model_year + 6/12 - calendar_year).